Extract Polyhedron class
